bot/requirements.txt
waitress==3.0.0
Flask==2.0.2
google.generativeai==0.3.2
Werkzeug==2.2.2
